Cast

+++ Unity/Assets/Scripts/Model/Generate/Client/Config/HeroConfig.cs
@@ -24,7 +24,7 @@ namespace ET
             MagicDef = _buf.ReadInt();
             AtkInterval = _buf.ReadFloat();
             AtkRange = _buf.ReadInt();
-            Speed = _buf.ReadInt();
+            Speed = _buf.ReadFloat();
             AOIRange = _buf.ReadInt();
             ResourceName = _buf.ReadString();
             ViewDistance = _buf.ReadLong();
@@ -80,7 +80,7 @@ namespace ET
         /// <summary>
         /// 移动速度
         /// </summary>
-        public readonly int Speed;
+        public readonly float Speed;
 
         /// <summary>
         /// 视野范围
